modules are generated but dependencies are missing
This is a long description of the current modification where I can detail anything.
Bundle contains all modules of the project.
Core module should define the domain objects and the relationships. In addition, it define the service interfaces.
Service module should depend on the core module and implement the service interfaces with corresponding (proper) business logic. But this module delegates the responsibility of storing data. So it defines DAO interfaces with CRUD methods.
dao-memory should depend on the service and implement the DAO interfaces with a certain technology.
controller should depend on the service and provides entry point for them.

Issue No: sorry we have no issue