core has been filled with book example
core module was extended with Author and Book classes which are represent our domain model. These model classes defines the data structure of an entity in our system.
In addition, core module defines the services which are contained in the api package. The services are defined by interfaces (or abstract classes). Definition should contain the following 5 parts: visibility, method name, parameters, exceptions and documentation (/** JavaDoc */).