springdateexample docker script added
Spring Data Example was created last week.
This example uses a MySQL database with the Sakila DVD Rental Template DB.
ShellScript creates a new docker network and run a mysql container.
When the MySQL Server is running and listening on 3306 (default port) then the SQL Schema and Data ar imported.
Host, IP, and passwords can be configured via variables.

Script was tested on Ubuntu. Import was checked with MySQL Workbench.