Calculator Web App
Basic Calculator Web Application was created with Spring Boot.
This example demonstrates how the controller, api and implementation are separated.
The api package defines the Calculator interface which represents the business logic.
The interface defines how the Calculator should be used but it contains no implementation.
The api also contains the CalculatorImpl which implements the Calculator interface and it has the business logic.

The config package contains the initialization of the required beans which is the Calculator in this case.
It creates a Calculator object whose dynamic type is CalculatorImpl.

The controller package defines the end point and maps the Business Logic to URL.