WebShop Demo is initialized
ProductQueryService is defined to list all Products.
This interface is implemented in ProductQueryServiceImp.java so definition and implementation are separated.
See: Design by Contract

ProductQueryServiceImpl.java still returns with a hard coded collection. But Controller is simplified. Note that this was related to the Single Responsibility Princible because Controller have less responsibilities now.

WebConfiguration.java was created to initialize the required Services DAO, etc.

Problems:
- ProductQueryServiceImpl.java still creates objects and we use no database.
- Product is still a bean.