webshop
Product.java was converted into an Entity.

ProductRepository.java was added as a CrudRepository. It already defines a lot of methods for data handling.

ProductQueryServiceImpl.java was modified to used ProductRepository.java and it expects it as constructor parameter.

Finally, WebConfiguration.java was modified in order to pass the ProductRepository.java to the Service initialization.