Init ShellScript is added.
This script was copied from another project.
The script expects a parameters start | stop.

When the container starts it runs mysql on a network with fix ip address. Then it loads the sakila schema and data SQL files to the server.

If the container stops then it is also removed from docker registry.

#3