Web Application is configured.
App.java was configured as SpringBootApplication.

WebConfiguration.java was added and it initialize the ActorSearchService bean which is an ActorSearchServiceImpl.java object. This Bean gets the ActorRepository as a parameter and pass it to the service implementation.

DataSource URL is required. It was defined in the application.properties file. You have to set these properties according to your own configuration.

#5